test/edu/uw/tcss/app/TruckTest.java
package edu.uw.tcss.app;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertFalse;
import static org.junit.jupiter.api.Assertions.assertNotEquals;
import static org.junit.jupiter.api.Assertions.assertTrue;

import edu.uw.tcss.model.Atv;
import edu.uw.tcss.model.Bicycle;
import edu.uw.tcss.model.Car;
import edu.uw.tcss.model.Direction;
import edu.uw.tcss.model.Human;
import edu.uw.tcss.model.Light;
import edu.uw.tcss.model.Taxi;
import edu.uw.tcss.model.Terrain;
import edu.uw.tcss.model.Truck;
import edu.uw.tcss.model.Vehicle;
import java.util.HashMap;
import java.util.Map;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

/**
 * Test class for the Truck class.
 *
 * @author Raiden H
 * @version Oct 23 2024
 */
public class TruckTest {
    /**
     * The number of times to repeat a test to have a high probability that all
     * random possibilities have been explored.
     */
    private static final int TRIES_FOR_RANDOMNESS = 10;
    /**
     * test map for testing chooseDirection().
     */
    private Map<Direction, Terrain> myTestMap = new HashMap<>();
    /**
     * test Truck object for testing Truck methods.
     */
    private Vehicle myTestTruck;
    @BeforeEach
    void setup() {
        myTestTruck = new Truck(0, 0, Direction.NORTH);
        myTestMap = new HashMap<>();
    }

    @Test
    void testGetImageFileName() {
        assertEquals("truck.gif", myTestTruck.getImageFileName(),
                "Living truck image filename is incorrect.");
    }
    @Test
    void testCanPassGrass() {
        final Terrain grass = Terrain.GRASS;
        assertFalse(myTestTruck.canPass(grass, Light.GREEN),
                "Truck should not be able to pass green-light grass.");
        assertFalse(myTestTruck.canPass(grass, Light.YELLOW),
                "Truck should not be able to pass yellow-light grass.");
        assertFalse(myTestTruck.canPass(grass, Light.RED),
                "Truck should not be able to pass red-light grass.");
    }
    @Test
    void testCanPassStreet() {
        final Terrain street = Terrain.STREET;
        assertTrue(myTestTruck.canPass(street, Light.GREEN),
                "Truck should be able to pass green-light street.");
        assertTrue(myTestTruck.canPass(street, Light.YELLOW),
                "Truck should be able to pass yellow-light street.");
        assertTrue(myTestTruck.canPass(street, Light.RED),
                "Truck should be able to pass red-light street.");
    }
    @Test
    void testCanPassLight() {
        final Terrain light = Terrain.LIGHT;
        assertTrue(myTestTruck.canPass(light, Light.GREEN),
                "Truck should be able to pass green-light light.");
        assertTrue(myTestTruck.canPass(light, Light.YELLOW),
                "Truck should be able to pass yellow-light light.");
        assertTrue(myTestTruck.canPass(light, Light.RED),
                "Truck should be able to pass red-light light.");
    }
    @Test
    void testCanPassWall() {
        final Terrain wall = Terrain.WALL;
        assertFalse(myTestTruck.canPass(wall, Light.GREEN),
                "Truck should not be able to pass green-light wall.");
        assertFalse(myTestTruck.canPass(wall, Light.YELLOW),
                "Truck should not be able to pass yellow-light wall.");
        assertFalse(myTestTruck.canPass(wall, Light.RED),
                "Truck should not be able to pass red-light wall.");
    }
    @Test
    void testCanPassTrail() {
        final Terrain trail = Terrain.TRAIL;
        assertFalse(myTestTruck.canPass(trail, Light.GREEN),
                "Truck should not be able to pass green-light trail.");
        assertFalse(myTestTruck.canPass(trail, Light.YELLOW),
                "Truck should not be able to pass yellow-light trail.");
        assertFalse(myTestTruck.canPass(trail, Light.RED),
                "Truck should not be able to pass red-light trail.");
    }
    @Test
    void testCanPassCrosswalk() {
        final Terrain crosswalk = Terrain.CROSSWALK;
        assertTrue(myTestTruck.canPass(crosswalk, Light.GREEN),
                "Truck should be able to pass green-light crosswalk.");
        assertTrue(myTestTruck.canPass(crosswalk, Light.YELLOW),
                "Truck should be able to pass yellow-light crosswalk.");
        assertFalse(myTestTruck.canPass(crosswalk, Light.RED),
                "Truck should not be able to pass red-light crosswalk.");
    }

    @Test
    void testChooseDirection() {
        //I know this is not pretty but the test cases aren't graded on warnings
        //and I didn't care enough to figure out how to prettify it.
        myTestMap.put(Direction.NORTH, Terrain.GRASS);
        myTestMap.put(Direction.SOUTH, Terrain.STREET);
        myTestMap.put(Direction.EAST, Terrain.WALL);
        myTestMap.put(Direction.WEST, Terrain.LIGHT);
        for (int i = 0; i < TRIES_FOR_RANDOMNESS; i++) {
            assertNotEquals(Direction.SOUTH, myTestTruck.chooseDirection(myTestMap),
                            "Truck should prefer any direction but backwards.");
        }

        myTestMap.clear();
        myTestMap.put(Direction.NORTH, Terrain.STREET);
        myTestMap.put(Direction.SOUTH, Terrain.STREET);
        myTestMap.put(Direction.EAST, Terrain.STREET);
        myTestMap.put(Direction.WEST, Terrain.STREET);
        for (int i = 0; i < TRIES_FOR_RANDOMNESS; i++) {
            assertNotEquals(Direction.SOUTH, myTestTruck.chooseDirection(myTestMap),
                    "Truck should prefer any direction but backwards.");
        }

        myTestMap.clear();
        myTestMap.put(Direction.NORTH, Terrain.GRASS);
        myTestMap.put(Direction.SOUTH, Terrain.STREET);
        myTestMap.put(Direction.EAST, Terrain.TRAIL);
        myTestMap.put(Direction.WEST, Terrain.WALL);
        assertEquals(Direction.SOUTH, myTestTruck.chooseDirection(myTestMap),
                "Truck should reverse if forced to.");

        myTestMap.clear();
        myTestMap.put(Direction.NORTH, Terrain.GRASS);
        myTestMap.put(Direction.SOUTH, Terrain.STREET);
        myTestMap.put(Direction.EAST, Terrain.CROSSWALK);
        myTestMap.put(Direction.WEST, Terrain.WALL);
        assertEquals(Direction.EAST, myTestTruck.chooseDirection(myTestMap),
                "Truck should take single exit over reversing.");
    }

    @Test
    void testCollide() {
        myTestTruck.collide(new Car(0, 0, Direction.NORTH));
        assertTrue(myTestTruck.isAlive(), "Truck should survive collision with Car");
        myTestTruck.collide(new Taxi(0, 0, Direction.NORTH));
        assertTrue(myTestTruck.isAlive(), "Truck should survive collision with Taxi");
        myTestTruck.collide(new Atv(0, 0, Direction.NORTH));
        assertTrue(myTestTruck.isAlive(), "Truck should survive collision with Atv");
        myTestTruck.collide(new Bicycle(0, 0, Direction.NORTH));
        assertTrue(myTestTruck.isAlive(), "Truck should survive collision with Bicycle");
        myTestTruck.collide(new Human(0, 0, Direction.NORTH));
        assertTrue(myTestTruck.isAlive(), "Truck should survive collision with Human");
    }

    @Test
    void testIsAlive() {
        assertTrue(myTestTruck.isAlive(),
                "Truck should always be alive");
    }
}